+++ src/components/burger-ingredients/burger-ingredients.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from './ingredient-details.module.css';
+import styles from './burger-ingredients.module.css';
 import { Tab, CurrencyIcon, Counter } from '@ya.praktikum/react-developer-burger-ui-components';
 import OrderDetails from '../order-details/order-details';
